# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standard API error response (RFC 7807-inspired).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable error description",
            "title": "Detail"
          },
          "error": {
            "description": "Machine-readable error code",
            "title": "Error",
            "type": "string"
          },
          "status_code": {
            "description": "HTTP status code",
            "title": "Status Code",
            "type": "integer"
          }
        },
        "required": [
          "error",
          "status_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_UsageResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsageResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "description": "Current page number (1-indexed)",
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "default": 25,
            "description": "Items per page",
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items matching the query",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "PaginatedResponse[UsageResponse]",
        "type": "object"
      },
...
      "UsageResponse": {
        "description": "Usage event response.\n\nSchema.org @type: UseAction\nMaps: tokens \u2192 result, timestamp \u2192 startTime",
        "properties": {
          "@context": {
            "default": "https://schema.org",
            "title": "@Context",
            "type": "string"
          },
          "@type": {
            "default": "UseAction",
            "title": "@Type",
            "type": "string"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "User ID (Schema.org: agent)",
            "title": "Agent"
          },
          "endpoint": {
            "default": "",
            "title": "Endpoint",
            "type": "string"
          },
          "event_type": {
            "description": "api_call, chat_inference, on_device_sync",
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "description": "Event document ID",
            "title": "Id",
            "type": "string"
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "API key ID (Schema.org: instrument)",
            "title": "Instrument"
          },
          "latency_ms": {
            "default": 0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "object": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model ID (Schema.org: object)",
            "title": "Object"
          },
          "region": {
            "default": "jnb",
            "title": "Region",
            "type": "string"
          },
          "start_time": {
            "description": "Event timestamp (Schema.org: startTime)",
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "tokens_input": {
            "default": 0,
            "title": "Tokens Input",
            "type": "integer"
          },
          "tokens_output": {
            "default": 0,
            "title": "Tokens Output",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "event_type",
          "start_time"
        ],
        "title": "UsageResponse",
        "type": "object"
      },
      "UsageSummaryResponse": {
        "description": "Aggregated usage summary for a time period.",
        "properties": {
          "average_latency_ms": {
            "default": 0.0,
            "title": "Average Latency Ms",
            "type": "number"
          },
          "error_count": {
            "default": 0,
            "title": "Error Count",
            "type": "integer"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "rate_limited_count": {
            "default": 0,
            "title": "Rate Limited Count",
            "type": "integer"
          },
          "success_count": {
            "default": 0,
            "title": "Success Count",
            "type": "integer"
          },
          "total_requests": {
            "default": 0,
            "title": "Total Requests",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          },
          "total_tokens_input": {
            "default": 0,
            "title": "Total Tokens Input",
            "type": "integer"
          },
          "total_tokens_output": {
            "default": 0,
            "title": "Total Tokens Output",
            "type": "integer"
          }
        },
        "required": [
          "period_start",
          "period_end"
        ],
        "title": "UsageSummaryResponse",
        "type": "object"
      },
      "UserPreferencesResponse": {
        "description": "User preferences (public projection).",
        "properties": {
          "default_model_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model Slug"
          },
          "language": {
            "description": "ISO 639-1 preferred language code",
            "title": "Language",
            "type": "string"
          },
          "notifications_enabled": {
            "title": "Notifications Enabled",
            "type": "boolean"
          },
          "theme": {
            "description": "UI theme: light, dark, or system",
            "title": "Theme",
            "type": "string"
          }
        },
        "required": [
          "language",
          "theme",
          "notifications_enabled"
        ],
        "title": "UserPreferencesResponse",
        "type": "object"
      },
      "UserResponse": {
        "description": "User profile response.\n\nSchema.org @type: Person\nMaps: display_name \u2192 name, avatar_url \u2192 image, email \u2192 email",
        "properties": {
          "@context": {
            "default": "https://schema.org",
            "title": "@Context",
            "type": "string"
          },
          "@type": {
            "default": "Person",
            "title": "@Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "description": "Email address (Schema.org: email)",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "description": "User identifier",
            "title": "Id",
            "type": "string"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Avatar URL (Schema.org: image)",
            "title": "Image"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_active_at": {
            "format": "date-time",
            "title": "Last Active At",
            "type": "string"
          },
          "name": {
            "description": "Display name (Schema.org: name)",
            "title": "Name",
            "type": "string"
          },
          "organization_ids": {
            "items": {
              "type": "string"
            },
            "title": "Organization Ids",
            "type": "array"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferencesResponse"
          },
          "role": {
            "description": "User role: consumer, developer, admin",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "preferences",
          "is_active",
          "last_active_at",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserUpdateRequest": {
        "description": "Update user profile request.",
        "properties": {
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New avatar URL",
            "title": "Image"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New display name",
            "title": "Name"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserPreferencesResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UserUpdateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "api@shamwari.ai",
      "name": "Nyuchi Web Services",
      "url": "https://shamwari.ai/"
    },
    "description": "Africa's open source AI platform \u2014 API backend.\n\nShamwari AI provides AI inference, model management, and developer APIs purpose-built for African languages and contexts. Built by Nyuchi Africa.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Shamwari AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/usage": {
      "get": {
        "description": "Returns paginated API usage events with Schema.org UseAction structured data. Filter by date range, model, or API key.",
        "operationId": "list_usage_v1_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "api_key_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key Id"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UsageResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List usage events",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/summary": {
      "get": {
        "description": "Returns aggregated usage metrics for the requested rolling window. Pass `range=7d|30d|90d`. Defaults to 30 days.",
        "operationId": "get_usage_summary_v1_usage_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "30d",
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get usage summary",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/users/me": {
      "get": {
        "description": "Returns the authenticated user's profile with Schema.org Person structured data. Includes preferences, role, and organization memberships.",
        "operationId": "get_current_user_profile_v1_users_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Update the authenticated user's display name, avatar, or preferences.",
        "operationId": "update_current_user_profile_v1_users_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid update parameters"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update current user profile",
        "tags": [
          "users"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.shamwari.ai"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8000"
    }
  ],
...
}